fix test type error

+++ examples/server/src/modules/todo/tests/Todo.service.test.ts
@@ -23,31 +23,44 @@ describe('Todo.service test', () => {
       Array(100)
         .fill('todo')
         .map((i, idx) =>
-          todoService.createTodo({
-            title: `${i}-${idx}`,
-            status: TodoStatus.todo,
-          })
+          todoService.createTodo(
+            {
+              title: `${i}-${idx}`,
+              status: TodoStatus.todo,
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         )
     );
   });
   after('delete todos', async () => {
-    const { data } = await todoService.queryTodo({});
-    return Promise.all(data.map((i) => todoService.deleteTodo({ id: i.id } as any)));
+    const { data } = await todoService.queryTodo({}, { type: 'admin', ip: '', lang: '' });
+    return Promise.all(
+      data.map((i) =>
+        todoService.deleteTodo({ id: i.id } as any, { type: 'admin', ip: '', lang: '' })
+      )
+    );
   });
 
   describe('todoService.createTodo and todoService.deleteTo', () => {
     it('todoService.queryTodo should success', async () => {
-      expect((await todoService.queryTodo({})).data.length).to.equal(100);
+      expect(
+        (await todoService.queryTodo({}, { type: 'admin', ip: '', lang: '' })).data.length
+      ).to.equal(100);
     });
 
     it('todoService.deleteTodo with `id` param should success', async () => {
-      await todoService.deleteTodo({ id: 1 } as any);
-      expect((await todoService.queryTodo({})).data.length).to.equal(99);
+      await todoService.deleteTodo({ id: 1 } as any, { type: 'admin', ip: '', lang: '' });
+      expect(
+        (await todoService.queryTodo({}, { type: 'admin', ip: '', lang: '' })).data.length
+      ).to.equal(99);
     });
 
     it('todoService.deleteTodo with `IDs` param should success', async () => {
-      await todoService.deleteTodo({ IDs: [2, 3] } as any);
-      expect((await todoService.queryTodo({})).data.length).to.equal(97);
+      await todoService.deleteTodo({ IDs: [2, 3] } as any, { type: 'admin', ip: '', lang: '' });
+      expect(
+        (await todoService.queryTodo({}, { type: 'admin', ip: '', lang: '' })).data.length
+      ).to.equal(97);
     });
   });
 
@@ -55,10 +68,13 @@ describe('Todo.service test', () => {
     it('queryTodo paginate should success', async () => {
       expect(
         (
-          await todoService.queryTodo({
-            perPage: 20,
-            page: 1,
-          })
+          await todoService.queryTodo(
+            {
+              perPage: 20,
+              page: 1,
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(20);
     });
@@ -66,9 +82,12 @@ describe('Todo.service test', () => {
     it('queryTodo with keyword `todo` should success', async () => {
       expect(
         (
-          await todoService.queryTodo({
-            keyword: 'todo',
-          })
+          await todoService.queryTodo(
+            {
+              keyword: 'todo',
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(97);
     });
@@ -76,9 +95,12 @@ describe('Todo.service test', () => {
     it('queryTodo with keyword `11` should success', async () => {
       expect(
         (
-          await todoService.queryTodo({
-            keyword: '11',
-          })
+          await todoService.queryTodo(
+            {
+              keyword: '11',
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(1);
     });
@@ -86,18 +108,24 @@ describe('Todo.service test', () => {
 
   describe('todoService.queryTodoByCursor', () => {
     it('queryTodoByCursor with keyword `todo` should success', async () => {
-      const result = await todoService.queryTodoByCursor({
-        keyword: 'todo',
-      });
+      const result = await todoService.queryTodoByCursor(
+        {
+          keyword: 'todo',
+        },
+        { type: 'admin', ip: '', lang: '' }
+      );
       expect(result.data.length).to.equal(97);
     });
 
     it('queryTodoByCursor with keyword `11` should success', async () => {
       expect(
         (
-          await todoService.queryTodoByCursor({
-            keyword: '99',
-          })
+          await todoService.queryTodoByCursor(
+            {
+              keyword: '99',
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(1);
     });
@@ -105,9 +133,12 @@ describe('Todo.service test', () => {
     it('queryTodoByCursor `perPage` 20 paginate should success', async () => {
       expect(
         (
-          await todoService.queryTodoByCursor({
-            perPage: 20,
-          })
+          await todoService.queryTodoByCursor(
+            {
+              perPage: 20,
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(20);
     });
@@ -115,33 +146,45 @@ describe('Todo.service test', () => {
     it('queryTodoByCursor paginate `perPage` 10 should success', async () => {
       expect(
         (
-          await todoService.queryTodoByCursor({
-            perPage: 10,
-          })
+          await todoService.queryTodoByCursor(
+            {
+              perPage: 10,
+            },
+            { type: 'admin', ip: '', lang: '' }
+          )
         ).data.length
       ).to.equal(10);
     });
 
     it('queryTodoByCursor paginate `beforeCursor` and `afterCursor`', async () => {
-      const result = await todoService.queryTodoByCursor({
-        perPage: 1,
-      });
+      const result = await todoService.queryTodoByCursor(
+        {
+          perPage: 1,
+        },
+        { type: 'admin', ip: '', lang: '' }
+      );
 
       expect(result.beforeCursor).to.equal(null);
       expect(typeof result.afterCursor).to.equal('string');
       expect(result.data.length).to.equal(1);
 
-      const result2 = await todoService.queryTodoByCursor({
-        perPage: 2,
-        afterCursor: result.afterCursor,
-      });
+      const result2 = await todoService.queryTodoByCursor(
+        {
+          perPage: 2,
+          afterCursor: result.afterCursor,
+        },
+        { type: 'admin', ip: '', lang: '' }
+      );
 
       expect(result2.data.length).to.equal(2);
       expect(result2.beforeCursor).not.to.equal(result.afterCursor);
 
-      const result3 = await todoService.queryTodoByCursor({
-        beforeCursor: result2.afterCursor,
-      });
+      const result3 = await todoService.queryTodoByCursor(
+        {
+          beforeCursor: result2.afterCursor,
+        },
+        { type: 'admin', ip: '', lang: '' }
+      );
 
       console.log(
         result.beforeCursor,
